// src/client-comms/types/MessageType.ts
export enum MessageType {
  INIT = 'INIT',
  DISCONNECT = 'DISCONNECT',
  BLACKLIST = 'BLACKLIST',
  REGION_REQUEST = 'REGION_REQUEST',
  REGION_REQUEST_RESPONSE = 'REGION_REQUEST_RESPONSE',
  SCAN_RESULT = 'SCAN_RESULT',
  UPDATE_AVAILABLE = 'UPDATE_AVAILABLE',
  DIMENSION_COMPLETE = 'DIMENSION_COMPLETE',
  REGION_COMPLETE = 'REGION_COMPLETE',
}